Implement authorization

+++ src/api/notes/handler.js
@@ -6,8 +6,13 @@ class NotesHandler {
 
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload)
-    const { title = undefined, body, tags } = request.payload
-    const noteId = await this._service.addNote({ title, body, tags })
+    const { title = 'untitled', body, tags } = request.payload
+    const { id: credentialId } = request.auth.credentials
+
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialId,
+    })
+
     const response = h.response({
       status: 'success',
       message: 'Catatan berhasil ditambahkan',
@@ -19,8 +24,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials
+    const notes = await this._service.getNotes(credentialId)
 
     return {
       status: 'success',
@@ -32,7 +38,10 @@ class NotesHandler {
 
   async getNoteByIdHandler(request) {
     const { id } = request.params;
-    const note = await this._service.getNoteById(id);
+    const { id: credentialId } = request.auth.credentials
+
+    await this._service.verifyNoteOwner(id, credentialId)
+    const note = await this._service.getNoteById(id, credentialId);
 
     return {
       status: 'success',
@@ -45,7 +54,9 @@ class NotesHandler {
   async putNoteByIdHandler(request) {
     this._validator.validateNotePayload(request.payload)
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials
 
+    await this._service.verifyNoteOwner(id, credentialId)
     await this._service.putNoteById(id, request.payload);
 
     return {
@@ -56,8 +67,11 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials
 
+    await this._service.verifyNoteOwner(id, credentialId)
     await this._service.deleteNoteById(id);
+
     return {
       status: 'success',
       message: 'Catatan berhasil dihapus',
